backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.1
hiredis==2.3.2

# Task Queue
celery==5.3.6
flower==2.0.1

# HTTP Requests
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Web Scraping
beautifulsoup4==4.12.3
selenium==4.17.2
lxml==5.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2

# AI & NLP
openai==1.10.0
transformers==4.37.0
torch==2.1.2

# Financial Data APIs
yfinance==0.2.35
alpha-vantage==2.3.1
finnhub-python==2.4.19

# News APIs
newsapi-python==0.2.7

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytz==2023.3.post1

# Monitoring & Logging
sentry-sdk==1.40.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# CORS
python-cors==1.0.0

